+++ scr/Operaciones.java
@@ -2,19 +2,19 @@ import java.util.HashSet;
 import java.util.Set;
 
 public class Operaciones {
-    public static int sumar(int num1, int num2) {
+    public static float sumar(float num1, float num2) {
         return num1 + num2;
     }
 
-    public static int restar(int num1, int num2) {
+    public static float restar(float num1, float num2) {
         return num1 - num2;
     }
 
-    public static int multiplicar(int num1, int num2) {
+    public static float multiplicar(float num1, float num2) {
         return num1 * num2;
     }
 
-    public static int dividir(int num1, int num2) {
+    public static float dividir(float num1, float num2) {
         return num1 / num2;
     }
 }
